functions/setup-reddit-final/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.13.0